CorrchoiceQualityBackend/requirements/common.txt
asgiref==3.8.1
djangorestframework==3.15.1
Django==4.2.13
mysqlclient==2.1.1
redis==5.0.5
sqlparse==0.5.0
PyYAML==6.0.1
typing-extensions==4.11.0
tzdata==2024.1